+++ src/rules.ts
@@ -1,5 +1,4 @@
 import { clues, Rule } from "./clues.js";
-import emojiRegex from "emoji-regex";
 import { Alt } from "./index.js";
 
 export function createSuggestion(ruleName: string, value?: string | number) {
@@ -29,24 +28,4 @@ export function checkClue(alt: Alt) {
     arr = [...arr, ...filterClues({ ruleName: item, clue: clue, alt })];
     return arr;
   }, []);
-}
-
-export function checkLength(alt: Alt) {
-  return alt.length > 125 ? [createSuggestion("charLength", alt.length)] : [];
-}
-
-export function checkOnlySpace(alt: Alt) {
-  return alt == " " ? [createSuggestion("notOnlySpace")] : [];
-}
-
-export function checkPunctuation(alt: Alt) {
-  return !/[.!?]$/.test(alt) && alt.length > 1
-    ? [createSuggestion("endPunctuation")]
-    : [];
-}
-
-export function checkEmoji(alt: Alt) {
-  const regex = emojiRegex();
-  const match = regex.exec(alt);
-  return match ? [createSuggestion("avoidEmoji", match[0])] : [];
 }
