+++ src/rules/__tests__/exact-match.test.ts
@@ -1,0 +1,9 @@
+import { checkClue } from "../../rules";
+
+test("exactMatch", () => {
+  expect(checkClue("logo")).toMatchInlineSnapshot(`
+      Array [
+        "Alt text should not be \\"logo\\" (https://tinyurl.com/yxwc2hof).",
+      ]
+    `);
+});
